+++ SpendingTrackerAPI/Services/IncomeCategoryService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SpendingTrackerAPI.Database;
 using SpendingTrackerAPI.Entities;
+using SpendingTrackerAPI.Exceptions;
 
 namespace SpendingTrackerAPI.Services;
 
@@ -25,6 +26,7 @@ public class IncomeCategoryService : IIncomeCategoryService
     public IncomeCategory GetById(int id)
     {
         var category = _context.IncomeCategories.FirstOrDefault(x => x.Id == id);
+        if (category == null) throw new NotFoundException("Not found!");
         return category;
     }
     
@@ -48,6 +50,7 @@ public class IncomeCategoryService : IIncomeCategoryService
     public void Update(int id ,IncomeCategory incomeCategory)
     {
         var dbCategory = _context.IncomeCategories.FirstOrDefault(c => c.Id == id);
+        if (dbCategory == null) throw new NotFoundException("Not found!");
         dbCategory.Name = incomeCategory.Name;
         _context.SaveChanges();
     }
@@ -55,6 +58,7 @@ public class IncomeCategoryService : IIncomeCategoryService
     public void Remove(int id)
     {
         var category = _context.IncomeCategories.FirstOrDefault(c => c.Id == id);
+        if (category == null) throw new NotFoundException("Not found!");
         _context.IncomeCategories.Remove(category);
         _context.SaveChanges();
     }
